Add API to clear the screen
Changed "draw" to "graphics" to make a little more sense and be
consistent with assets.

A clearScreen function allows for animation - main.lua shows an example
of how that works. Note that it doesn't matter where clearScreen is put
(so long as it's called once per frame). In other words, all draw calls
are batched together and pushed to the screen all at once.